requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai-whisper==1.3.1
pydub==0.25.1
python-multipart==0.0.9
python-dotenv==1.0.1
ffmpeg-python==0.2.0
